requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
faiss-cpu>=1.7.4
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5
numpy>=1.21.0